CoursePage split, finally
